backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "motif-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-cloud-firestore>=2.16.0",
    "google-cloud-storage>=2.16.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "Pillow>=10.0",
    "numpy>=1.26",
    "rembg[cpu]>=2.0",
    "sentence-transformers>=3.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]
